Refactor stats extraction into its own module
